maintain the same size on both 32/64 bit architecture
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@175698 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/internal/XSerializeEngine.hpp
@@ -56,6 +56,9 @@
 
 /*
  * $Log$
+ * Revision 1.12  2004/01/16 21:55:18  peiyongz
+ * maintain the same size on both 32/64 bit architecture
+ *
  * Revision 1.11  2004/01/15 23:42:32  peiyongz
  * proper allignment for built-in datatype read/write
  *
@@ -124,7 +127,7 @@ public:
     static int defaultDataLen;
 
 
-    typedef unsigned long   XSerializedObjectId_t;
+    typedef unsigned int   XSerializedObjectId_t;
 
     /***
       *
